Location working correctly
