+++ app/src/main/java/com/example/scansaga/Views/ScanAndGo.java
@@ -2,23 +2,41 @@ package com.example.scansaga.Views;
 
 import static com.example.scansaga.Model.MainActivity.token;
 
+import android.Manifest;
+import android.annotation.SuppressLint;
+import android.content.Context;
 import android.content.Intent;
+import android.content.pm.PackageManager;
 import android.graphics.Bitmap;
+import android.location.Location;
+import android.location.LocationManager;
 import android.net.Uri;
 import android.os.Bundle;
+import android.os.Looper;
 import android.provider.MediaStore;
 import android.provider.Settings;
+import android.support.annotation.NonNull;
 import android.util.Log;
 import android.view.View;
 import android.widget.Button;
 import android.widget.ImageButton;
+import android.widget.Switch;
 import android.widget.Toast;
+import android.widget.ToggleButton;
 
 import androidx.annotation.Nullable;
 import androidx.appcompat.app.AppCompatActivity;
+import androidx.core.app.ActivityCompat;
 
 import com.example.scansaga.Model.GeoLocationManager;
 import com.example.scansaga.R;
+import com.google.android.gms.location.FusedLocationProviderClient;
+import com.google.android.gms.location.LocationCallback;
+import com.google.android.gms.location.LocationRequest;
+import com.google.android.gms.location.LocationResult;
+import com.google.android.gms.location.LocationServices;
+import com.google.android.gms.tasks.OnCompleteListener;
+import com.google.android.gms.tasks.Task;
 import com.google.firebase.firestore.DocumentSnapshot;
 import com.google.firebase.firestore.FieldValue;
 import com.google.firebase.firestore.FirebaseFirestore;
@@ -43,15 +61,24 @@ import java.util.Map;
 public class ScanAndGo extends AppCompatActivity {
     FirebaseFirestore db = FirebaseFirestore.getInstance();
     private static final int REQUEST_CODE_GALLERY = 1; // Add this line
+    FusedLocationProviderClient mFusedLocationClient;
+    int PERMISSION_ID = 44;
+    private String deviceId;
+    private ToggleButton geoLocation;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.camera_scan_page);
+        deviceId = Settings.Secure.getString(getContentResolver(), Settings.Secure.ANDROID_ID);
+        mFusedLocationClient = LocationServices.getFusedLocationProviderClient(this);
 
         ImageButton takePictureButton = findViewById(R.id.take_picture_button);
         ImageButton getPhotoButton = findViewById(R.id.get_photo_from_gallery_button);
 
+        // Inside onCreate or appropriately initialize this if using after onCreate
+        Switch geoLocationSwitch = findViewById(R.id.geo_location_switch);
+        geoLocationSwitch.setOnCheckedChangeListener((buttonView, isChecked) -> handleGeolocationToggle(isChecked));
 
         takePictureButton.setOnClickListener(new View.OnClickListener() {
             @Override
@@ -71,6 +98,16 @@ public class ScanAndGo extends AppCompatActivity {
                 Intent intent = new Intent(Intent.ACTION_PICK);
                 intent.setType("image/*");
                 startActivityForResult(intent, REQUEST_CODE_GALLERY);
+            }
+        });
+
+        geoLocationSwitch.setOnCheckedChangeListener((buttonView, isChecked) -> {
+            if (isChecked) {
+                // User turned on the geolocation toggle
+                handleGeolocationToggle(true);
+            } else {
+                // User turned off the geolocation toggle
+                handleGeolocationToggle(false);
             }
         });
 
@@ -171,7 +208,10 @@ public class ScanAndGo extends AppCompatActivity {
                     db.collection("events").document(url).update("attendeeCheckInCounts", attendeeCheckInCounts).addOnSuccessListener(aVoid -> {
                         Log.d("ScanAndGo", "Updated attendee check-in count");
                         // Redirect based on if it's a first check-in or a repeat
-                        redirectToCheckinResultPage(currentCount > 0 ? "Checked In Again" : "First Check-In", true);
+                        // Also check if they have location enabled. Add their location to db if it is
+                        if (checkPermissions()){
+                            uploadLocationAndCheckInAttendee(url,deviceId,GeoLocationManager.getLatitude(),GeoLocationManager.getLongitude());
+                        } else {redirectToCheckinResultPage(currentCount > 0 ? "Checked In Again" : "First Check-In", true);}
                     });
                 } else {
                     // No attendees have checked in yet, so add this user as the first
@@ -184,11 +224,14 @@ public class ScanAndGo extends AppCompatActivity {
                             });
 
                     db.collection("events").document(url).update("attendeeCheckInCounts", firstCheckIn);
-                    redirectToCheckinResultPage("First Check-In", true);
+                    if (checkPermissions()){ // need to check to see if location is permitted or not here somehow
+                        uploadLocationAndCheckInAttendee(url,deviceId,GeoLocationManager.getLatitude(),GeoLocationManager.getLongitude());
+                    } else {redirectToCheckinResultPage("First Check-In", true);}
+
                 }
             } else {
                 Log.d("ScanAndGo", "Couldn't check into event");
-                redirectToCheckinResultPage("Error checking event", false);
+                redirectToCheckinResultPage("Couldn't Find Event", false);
             }
         });
     }
@@ -209,18 +252,11 @@ public class ScanAndGo extends AppCompatActivity {
         locationData.put("deviceId", deviceId);
         locationData.put("latitude", latitude);
         locationData.put("longitude", longitude);
-        // Still need to add user to checked in list, so also execute the checkInWithoutLocation
-        checkInAttendeeWithoutLocation(url,deviceId);
 
         db.collection("events").document(url).collection("locationOfCheckedInUsers").document(deviceId)
                 .set(locationData)
                 .addOnSuccessListener(s -> redirectToCheckinResultPage("Location recorded & checked in successfully", true))
                 .addOnFailureListener(f -> redirectToCheckinResultPage("Error saving location data", false));
-
-        db.collection("events").document(url).update("signedUpAttendeeTokens", FieldValue.arrayUnion(token))
-                .addOnSuccessListener(v -> {
-                    Log.d("TOKEN", "Token successfully added in ScanAndGo");
-                });
     }
 
     /**
@@ -238,23 +274,129 @@ public class ScanAndGo extends AppCompatActivity {
     }
 
 
-    /**
-     * Adds the user to the list of checked-in attendees without location data.
-     * This method is a fallback for when location permissions are not granted or location data is not available.
-     *
-     * @param url      The URL identifying the event document in Firestore.
-     * @param deviceId The device ID of the user's device, used as a unique identifier.
-     */
-    private void checkInAttendeeWithoutLocation(String url, String deviceId) {
-        db.collection("events").document(url).update("checkedInAttendees", FieldValue.arrayUnion(deviceId))
-                .addOnSuccessListener(s -> redirectToCheckinResultPage("Checked in successfully", true))
-                .addOnFailureListener(f -> redirectToCheckinResultPage("Error checking in", false));
+    @SuppressLint("MissingPermission")
+    private void getLastLocation() {
+        // check if permissions are given
+        if (checkPermissions()) {
 
-        db.collection("events").document(url).update("signedUpAttendeeTokens", FieldValue.arrayUnion(token))
-                .addOnSuccessListener(v -> {
-                    Log.d("TOKEN", "Token successfully added in ScanAndGo");
+            // check if location is enabled
+            if (isLocationEnabled()) {
+
+                // getting last
+                // location from
+                // FusedLocationClient
+                // object
+                mFusedLocationClient.getLastLocation().addOnCompleteListener(new OnCompleteListener<Location>() {
+                    @Override
+                    public void onComplete(@NonNull Task<Location> task) {
+                        Location location = task.getResult();
+                        if (location == null) {
+                            requestNewLocationData();
+                        } else {
+                            GeoLocationManager.setLatitude(location.getLatitude());
+                            GeoLocationManager.setLongitude(location.getLongitude());
+                        }
+                    }
                 });
+            } else {
+                Toast.makeText(this, "Please turn on" + " your location...", Toast.LENGTH_LONG).show();
+                Intent intent = new Intent(Settings.ACTION_LOCATION_SOURCE_SETTINGS);
+                startActivity(intent);
+            }
+        } else {
+            // if permissions aren't available,
+            // request for permissions
+            requestPermissions();
+        }
     }
 
+    @SuppressLint("MissingPermission")
+    private void requestNewLocationData() {
+
+        // Initializing LocationRequest
+        // object with appropriate methods
+        LocationRequest mLocationRequest = new LocationRequest();
+        mLocationRequest.setPriority(LocationRequest.PRIORITY_HIGH_ACCURACY);
+        mLocationRequest.setInterval(5);
+        mLocationRequest.setFastestInterval(0);
+        mLocationRequest.setNumUpdates(1);
+
+        // setting LocationRequest
+        // on FusedLocationClient
+        mFusedLocationClient = LocationServices.getFusedLocationProviderClient(this);
+        mFusedLocationClient.requestLocationUpdates(mLocationRequest, mLocationCallback, Looper.myLooper());
+    }
+
+    private LocationCallback mLocationCallback = new LocationCallback() {
+
+        @Override
+        public void onLocationResult(LocationResult locationResult) {
+            Location mLastLocation = locationResult.getLastLocation();
+            GeoLocationManager.setLatitude(mLastLocation.getLatitude());
+            GeoLocationManager.setLongitude(mLastLocation.getLongitude());
+        }
+    };
+
+    // method to check for permissions
+    private boolean checkPermissions() {
+        return ActivityCompat.checkSelfPermission(this, Manifest.permission.ACCESS_COARSE_LOCATION) == PackageManager.PERMISSION_GRANTED && ActivityCompat.checkSelfPermission(this, Manifest.permission.ACCESS_FINE_LOCATION) == PackageManager.PERMISSION_GRANTED;
+    }
+
+    // method to request for permissions
+    private void requestPermissions() {
+        ActivityCompat.requestPermissions(this, new String[]{
+                Manifest.permission.ACCESS_COARSE_LOCATION,
+                Manifest.permission.ACCESS_FINE_LOCATION}, PERMISSION_ID);
+    }
+
+    // method to check
+    // if location is enabled
+    private boolean isLocationEnabled() {
+        LocationManager locationManager = (LocationManager) getSystemService(Context.LOCATION_SERVICE);
+        return locationManager.isProviderEnabled(LocationManager.GPS_PROVIDER) || locationManager.isProviderEnabled(LocationManager.NETWORK_PROVIDER);
+    }
+
+    // If everything is alright then
+    @Override
+    public void
+    onRequestPermissionsResult(int requestCode, @NonNull String[] permissions, @NonNull int[] grantResults) {
+        super.onRequestPermissionsResult(requestCode, permissions, grantResults);
+
+        if (requestCode == PERMISSION_ID) {
+            if (grantResults.length > 0 && grantResults[0] == PackageManager.PERMISSION_GRANTED) {
+                getLastLocation();
+            }
+        }
+    }
+
+    @Override
+    protected void onResume() {
+        super.onResume();
+        updateGeoLocationToggleState();
+        if (checkPermissions()) {
+            getLastLocation();
+        }
+    }
+    private void updateGeoLocationToggleState() {
+        Switch geoLocation = findViewById(R.id.geo_location_switch);
+        geoLocation.setChecked(checkLocationPermissions() && isLocationEnabled());
+    }
+
+    // Now the checkPermissions method will be the source of truth for permissions
+    private boolean checkLocationPermissions() {
+        return ActivityCompat.checkSelfPermission(this, Manifest.permission.ACCESS_COARSE_LOCATION) == PackageManager.PERMISSION_GRANTED && ActivityCompat.checkSelfPermission(this, Manifest.permission.ACCESS_FINE_LOCATION) == PackageManager.PERMISSION_GRANTED;
+    }
+    // Directly use checkPermissions() instead of GeoLocationManager.checkPermissions()
+    private void handleGeolocationToggle(boolean isEnabled) {
+        if (isEnabled) {
+            if (checkPermissions()) {
+                getLastLocation();
+            } else {
+                requestPermissions();
+            }
+        } else {
+            Toast.makeText(this, "Geolocation is disabled", Toast.LENGTH_SHORT).show();
+        }
+    }
 }
