+++ app/src/main/res/layout/camera_scan_page.xml
@@ -6,9 +6,9 @@
     android:layout_height="match_parent">
 
     <TableLayout
-        android:layout_width="412dp"
-        android:layout_height="200dp"
-        android:padding="25dp"
+        android:layout_width="411dp"
+        android:layout_height="258dp"
+        android:padding="10dp"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintLeft_toLeftOf="parent"
         app:layout_constraintRight_toRightOf="parent"
@@ -23,20 +23,20 @@
                 android:layout_width="0dp"
                 android:layout_height="wrap_content"
                 android:layout_weight="1"
+                android:padding="10dp"
                 android:text="Scan QR Code"
                 android:textAlignment="center"
-                android:textStyle="bold"
-                android:padding="10dp"/>
+                android:textStyle="bold" />
 
             <TextView
                 android:id="@+id/get_qr_from_gallery_textview"
                 android:layout_width="0dp"
                 android:layout_height="wrap_content"
                 android:layout_weight="1"
+                android:padding="10dp"
                 android:text="Get QR From Gallery"
                 android:textAlignment="center"
                 android:textStyle="bold"
-                android:padding="10dp"
                 app:srcCompat="@drawable/add_photo_asset" />
         </TableRow>
 
@@ -49,21 +49,32 @@
                 android:layout_width="0dp"
                 android:layout_height="100dp"
                 android:layout_weight="1"
-                android:scaleType="fitCenter"
+                android:contentDescription="Scan QR with Camera"
                 android:padding="10dp"
-                app:srcCompat="@drawable/camera_image_asset"
-                android:contentDescription="Scan QR with Camera"/>
+                android:scaleType="fitCenter"
+                app:srcCompat="@drawable/camera_image_asset" />
 
             <ImageButton
                 android:id="@+id/get_photo_from_gallery_button"
                 android:layout_width="0dp"
                 android:layout_height="100dp"
                 android:layout_weight="1"
-                android:scaleType="fitCenter"
-                android:padding="10dp"
-                app:srcCompat="@drawable/add_photo_asset"
                 android:contentDescription="Find QR from gallery"
-                />
+                android:padding="10dp"
+                android:scaleType="fitCenter"
+                app:srcCompat="@drawable/add_photo_asset" />
         </TableRow>
+
+        <Switch
+            android:id="@+id/geo_location_switch"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_gravity="center"
+            android:forceHasOverlappingRendering="true"
+            android:foregroundGravity="center"
+            android:gravity="center"
+            android:switchTextAppearance="@style/TextAppearance.AppCompat.Display1"
+            android:text="@string/geo_location_label"
+            android:textAlignment="viewEnd" />
     </TableLayout>
 </androidx.constraintlayout.widget.ConstraintLayout>
