+++ app/src/main/res/layout/homepage.xml
@@ -108,17 +108,6 @@
             android:layout_marginBottom="2sp"
             android:text="My Events" />
 
-        <Switch
-            android:id="@+id/geo_location_switch"
-            android:layout_width="289dp"
-            android:layout_height="wrap_content"
-            android:layout_gravity="center"
-            android:forceHasOverlappingRendering="true"
-            android:foregroundGravity="center"
-            android:gravity="center|right"
-            android:switchTextAppearance="@style/TextAppearance.AppCompat.Display1"
-            android:text="@string/geo_location_label" />
-
     </LinearLayout>
